requirements.txt
py-cord==2.0.0b5
pyfiglet==0.8.post1
dismusic
python-dotenv==0.20.0